requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
bcrypt==4.0.1
email-validator==2.0.0
gunicorn==21.2.0
requests-oauthlib==1.3.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
Flask-Mail==0.10.0
google-cloud-resource-manager==1.10.4
google-cloud-bigquery==3.11.4
google-cloud-iam==2.12.1
google-cloud-service-usage==1.8.3
google-cloud-billing==1.12.1
